- separated train_eval_script to three separate files - output metrics as csv in eval.py
